requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
Pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
telegraph==2.2.0
jieba==0.42.1
python-dateutil==2.8.2
wordcloud==1.9.3 